chore: make edits after renaming
