+++ bases/renku_data_services/data_api/main.py
@@ -22,7 +22,7 @@ def create_app() -> Sanic:
         # specified below with the main_process_start decorator do not run.
         run_migrations_for_app("resource_pools", config.rp_repo)
         run_migrations_for_app("storage", config.rp_repo)
-        run_migrations_for_app("projects", config.project_authz)
+        run_migrations_for_app("authz", config.project_authz)
         config.rp_repo.initialize(config.default_resource_pool)
     app = register_all_handlers(app, config)
 
@@ -31,7 +31,7 @@ def create_app() -> Sanic:
         logger.info("running migrations")
         run_migrations_for_app("resource_pools", config.rp_repo)
         run_migrations_for_app("storage", config.rp_repo)
-        run_migrations_for_app("projects", config.project_authz)
+        run_migrations_for_app("authz", config.project_authz)
         config.rp_repo.initialize(config.default_resource_pool)
 
     @app.before_server_start
