+++ components/renku_data_services/authz_migrations/__init__.py
@@ -1,1 +1,1 @@
-"""Database migrations for project models."""
+"""Database migrations for authz models."""
