+++ components/renku_data_services/authz_migrations/env.py
@@ -1,6 +1,6 @@
 """Alembic setup and environment used for database migrations."""
 from renku_data_services.migrations.env import run_migrations
-from renku_data_services.project.orm import BaseORM
+from renku_data_services.authz.orm import BaseORM
 
 target_metadata = BaseORM.metadata
 run_migrations(target_metadata)
