+++ components/renku_data_services/authz_migrations/versions/d8e4a9411c0f_create_project_user_permissions.py
@@ -1,8 +1,8 @@
-"""create authz table
+"""create project-user permissions
 
-Revision ID: 5097da80da51
-Revises:
-Create Date: 2023-11-07 21:22:08.404004
+Revision ID: d8e4a9411c0f
+Revises: 
+Create Date: 2023-11-12 20:54:04.944865
 
 """
 from alembic import op
@@ -10,7 +10,7 @@ import sqlalchemy as sa
 
 
 # revision identifiers, used by Alembic.
-revision = "5097da80da51"
+revision = "d8e4a9411c0f"
 down_revision = None
 branch_labels = None
 depends_on = None
@@ -25,38 +25,32 @@ def upgrade() -> None:
         sa.Column("user_id", sa.String(length=36), nullable=True),
         sa.Column("id", sa.Integer(), nullable=False),
         sa.PrimaryKeyConstraint("id"),
-        schema="projects",
+        schema="authz",
     )
     op.create_index(
-        op.f("ix_projects_project_user_authz_access_level"),
+        op.f("ix_authz_project_user_authz_access_level"),
         "project_user_authz",
         ["access_level"],
         unique=False,
-        schema="projects",
+        schema="authz",
     )
     op.create_index(
-        op.f("ix_projects_project_user_authz_project_id"),
+        op.f("ix_authz_project_user_authz_project_id"),
         "project_user_authz",
         ["project_id"],
         unique=False,
-        schema="projects",
+        schema="authz",
     )
     op.create_index(
-        op.f("ix_projects_project_user_authz_user_id"),
-        "project_user_authz",
-        ["user_id"],
-        unique=False,
-        schema="projects",
+        op.f("ix_authz_project_user_authz_user_id"), "project_user_authz", ["user_id"], unique=False, schema="authz"
     )
     # ### end Alembic commands ###
 
 
 def downgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_projects_project_user_authz_user_id"), table_name="project_user_authz", schema="projects")
-    op.drop_index(op.f("ix_projects_project_user_authz_project_id"), table_name="project_user_authz", schema="projects")
-    op.drop_index(
-        op.f("ix_projects_project_user_authz_access_level"), table_name="project_user_authz", schema="projects"
-    )
-    op.drop_table("project_user_authz", schema="projects")
+    op.drop_index(op.f("ix_authz_project_user_authz_user_id"), table_name="project_user_authz", schema="authz")
+    op.drop_index(op.f("ix_authz_project_user_authz_project_id"), table_name="project_user_authz", schema="authz")
+    op.drop_index(op.f("ix_authz_project_user_authz_access_level"), table_name="project_user_authz", schema="authz")
+    op.drop_table("project_user_authz", schema="authz")
     # ### end Alembic commands ###
